+++ knf/validators/network/validators_test.go
@@ -45,6 +45,12 @@ const _CONFIG_DATA = `
 	test0:
 	test1: https://google.com
 	test2: google.com/abcd.php
+
+[mail]
+	test0:
+	test1: unknown
+	test2: unknown@domain
+	test3: unknown@domain.com
 `
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -81,7 +87,7 @@ func (s *ValidatorSuite) TestIPValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
-	c.Assert(errs[0].Error(), Equals, "300.0.400.5 is not a valid IP address")
+	c.Assert(errs[0].Error(), Equals, `"300.0.400.5" is not a valid IP address`)
 }
 
 func (s *ValidatorSuite) TestHasIPValidator(c *C) {
@@ -102,7 +108,7 @@ func (s *ValidatorSuite) TestHasIPValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
-	c.Assert(errs[0].Error(), Equals, "The system does not have an interface with the address 192.168.1.254")
+	c.Assert(errs[0].Error(), Equals, `The system does not have an interface with the address "192.168.1.254"`)
 }
 
 func (s *ValidatorSuite) TestPortValidator(c *C) {
@@ -124,8 +130,8 @@ func (s *ValidatorSuite) TestPortValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 2)
-	c.Assert(errs[0].Error(), Equals, "ABCD is not a valid port number")
-	c.Assert(errs[1].Error(), Equals, "78361 is not a valid port number")
+	c.Assert(errs[0].Error(), Equals, `"ABCD" is not a valid port number`)
+	c.Assert(errs[1].Error(), Equals, `"78361" is not a valid port number`)
 }
 
 func (s *ValidatorSuite) TestMACValidator(c *C) {
@@ -146,7 +152,7 @@ func (s *ValidatorSuite) TestMACValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
-	c.Assert(errs[0].Error(), Equals, "ABCD is not a valid MAC address: address ABCD: invalid MAC address")
+	c.Assert(errs[0].Error(), Equals, `"ABCD" is not a valid MAC address: address ABCD: invalid MAC address`)
 }
 
 func (s *ValidatorSuite) TestCIDRValidator(c *C) {
@@ -167,7 +173,7 @@ func (s *ValidatorSuite) TestCIDRValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
-	c.Assert(errs[0].Error(), Equals, "127.0.0.1/200 is not a valid CIDR address: invalid CIDR address: 127.0.0.1/200")
+	c.Assert(errs[0].Error(), Equals, `"127.0.0.1/200" is not a valid CIDR address: invalid CIDR address: 127.0.0.1/200`)
 }
 
 func (s *ValidatorSuite) TestURLValidator(c *C) {
@@ -188,7 +194,25 @@ func (s *ValidatorSuite) TestURLValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
-	c.Assert(errs[0].Error(), Equals, "google.com/abcd.php is not a valid URL address: parse \"google.com/abcd.php\": invalid URI for request")
+	c.Assert(errs[0].Error(), Equals, `"google.com/abcd.php" is not a valid URL address: parse "google.com/abcd.php": invalid URI for request`)
+}
+
+func (s *ValidatorSuite) TestMailValidator(c *C) {
+	configFile := createConfig(c, _CONFIG_DATA)
+
+	err := knf.Global(configFile)
+	c.Assert(err, IsNil)
+
+	errs := knf.Validate([]*knf.Validator{
+		{"mail:test0", Mail, nil},
+		{"mail:test1", Mail, nil},
+		{"mail:test2", Mail, nil},
+		{"mail:test3", Mail, nil},
+	})
+
+	c.Assert(errs, HasLen, 2)
+	c.Assert(errs[0].Error(), Equals, `"unknown" is not a valid email address`)
+	c.Assert(errs[1].Error(), Equals, `"unknown@domain" is not a valid email address`)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //